refactor(guardian): move the guard constructor registry into helpers/
`registry.ts` is a utility module — two exported functions over a
module-private Map, no class and no exposed state — so it belongs beside
the package's other helpers rather than among the guard classes.

Follows the pattern `helpers/` already sets: re-exported from
`helpers/mod.ts` like every other helper (still not public API —
`deno.json` exposes only `.`, `./guards`, `./types`, `./errors`), guards
import it through that barrel per the cross-folder import rule, and the
test moves to `tests/helpers/registry.test.ts` importing via the barrel
like its neighbours.

Merged ArrayGuardian's two `../helpers/mod.ts` imports into one, which
the move would otherwise have left duplicated.

The DAG proof is now stronger: the whole package's value-import graph is
acyclic, and `helpers/registry.ts` value-imports only `errors/Base.ts`,
so it remains a sink and every reference to a guard stays `import type`.
