+++ packages/guardian/guards/NumberGuardian.ts
@@ -18,7 +18,7 @@ import type { GuardianMetaData, GuardianTransform } from '../types/mod.ts';
 import type { DateGuardian } from './DateGuardian.ts';
 import type { StringGuardian } from './StringGuardian.ts';
 import type { BigIntGuardian } from './BigIntGuardian.ts';
-import { registerGuardian, resolveGuardian } from './registry.ts';
+import { registerGuardian, resolveGuardian } from '../helpers/mod.ts';
 
 /**
  * Number validator. Coerces numeric strings, bigints, booleans, and
