+++ packages/guardian/helpers/registry.ts
@@ -4,37 +4,39 @@
  *
  * A transition like `NumberGuardian.toString()` has to produce a
  * {@link StringGuardian}, and `StringGuardian.toNumber()` has to produce
- * a {@link NumberGuardian}. Importing the sibling class as a VALUE for
- * that makes the guard modules mutually dependent — a legal ESM cycle
- * that native module evaluation handles fine, but that DEADLOCKS once a
- * bundler lowers module initialisers to async functions (which
- * esbuild / Rollup do for the whole graph as soon as anything in it uses
- * top-level await).
+ * a {@link NumberGuardian}. Having each guard import its sibling as a
+ * VALUE for that makes the guard modules mutually dependent — a legal
+ * ESM cycle that native module evaluation handles fine, but that
+ * DEADLOCKS once a bundler lowers module initialisers to async functions
+ * (which esbuild / Rollup do for the whole graph as soon as anything in
+ * it uses top-level await).
  *
- * The fix is to keep every sibling reference type-only — `import type`
+ * The fix is to keep every reference to a guard type-only — `import type`
  * is erased at compile time and cannot create a runtime edge — and to
  * look the CONSTRUCTOR up here instead. Each guard module registers
  * itself immediately after its class declaration; transition methods
  * resolve the target lazily, inside the method body, long after every
  * module in the barrel has finished evaluating.
  *
- * This module deliberately imports nothing from its siblings at runtime,
- * so the value-import graph under `guards/` stays acyclic.
+ * Living under `helpers/` is what keeps that true: nothing here imports a
+ * guard at runtime, so this module is a sink in the value-import graph
+ * and `guards/` stays acyclic. The only runtime dependency is the
+ * package error class.
  *
  * @module
  */
 
 import { GuardianError } from '../errors/Base.ts';
-import type { BigIntGuardian } from './BigIntGuardian.ts';
-import type { DateGuardian } from './DateGuardian.ts';
-import type { NumberGuardian } from './NumberGuardian.ts';
-import type { StringGuardian } from './StringGuardian.ts';
+import type { BigIntGuardian } from '../guards/BigIntGuardian.ts';
+import type { DateGuardian } from '../guards/DateGuardian.ts';
+import type { NumberGuardian } from '../guards/NumberGuardian.ts';
+import type { StringGuardian } from '../guards/StringGuardian.ts';
 
 /**
  * The guard classes reachable through the registry, keyed by the name
  * transition methods use. Only classes that are the TARGET of a
  * cross-guard transition appear here — a guard that merely transitions
- * away (e.g. {@link BooleanGuardian}) never needs to be resolved.
+ * away (e.g. `BooleanGuardian`) never needs to be resolved.
  *
  * The values are `typeof` the class, i.e. the constructor side, so a
  * resolved entry is assignable to `process()`'s `constructor` parameter
